chore: move navigation files
